+++ ext/monitor/src/test/java/app/TestModule.java
@@ -1,7 +1,10 @@
 package app;
 
+import app.monitor.kube.KubeClient;
 import core.framework.test.module.AbstractTestModule;
 import core.framework.util.ClasspathResources;
+
+import static org.mockito.Mockito.mock;
 
 public class TestModule extends AbstractTestModule {
     @Override
@@ -9,6 +12,7 @@ public class TestModule extends AbstractTestModule {
         System.setProperty("app.alert.config", ClasspathResources.text("alert.json"));
         System.setProperty("app.monitor.config", ClasspathResources.text("monitor.json"));
 
+        overrideBinding(KubeClient.class, mock(KubeClient.class));
         load(new MonitorApp());
     }
 }